Move closures to a separate module
This will allow adding other helpers for closures here as well without
cluttering lib.rs.
